Devel

+++ continuousflex/protocols/protocol_nma_alignment.py
@@ -31,17 +31,17 @@
 
 from os.path import basename
 
-from pyworkflow.em.convert.atom_struct import cifToPdb
+from pwem.convert.atom_struct import cifToPdb
 from pyworkflow.utils import replaceBaseExt
 
 from pyworkflow.utils import isPower2, getListFromRangeString
 from pyworkflow.utils.path import copyFile, cleanPath 
 import pyworkflow.protocol.params as params
-from pyworkflow.em.protocol import ProtAnalysis3D
+from pwem.protocols import ProtAnalysis3D
 
 from pyworkflow.protocol.params import NumericRangeParam
-import pyworkflow.em as em
-import pyworkflow.em.metadata as md
+import pwem as em
+import pwem.emlib.metadata as md
 
 from xmipp3.base import XmippMdRow
 from xmipp3.convert import (writeSetOfParticles, xmippToLocation,
@@ -188,7 +188,7 @@ class FlexProtAlignmentNMA(ProtAnalysis3D):
             # Conside the index is the id in the input set
             particle = inputSet[index]
             mdImgs.setValue(md.MDL_IMAGE, getImageLocation(particle), objId)
-            mdImgs.setValue(md.MDL_ITEM_ID, long(particle.getObjId()), objId)
+            mdImgs.setValue(md.MDL_ITEM_ID, int(particle.getObjId()), objId)
         mdImgs.write(self.imgsFn)
         
     def performNmaStep(self, atomsFn, modesFn):
@@ -220,7 +220,7 @@ class FlexProtAlignmentNMA(ProtAnalysis3D):
             # Conside the index is the id in the input set
             particle = inputSet[index]
             mdImgs.setValue(md.MDL_IMAGE, getImageLocation(particle), objId)
-            mdImgs.setValue(md.MDL_ITEM_ID, long(particle.getObjId()), objId)
+            mdImgs.setValue(md.MDL_ITEM_ID, int(particle.getObjId()), objId)
         mdImgs.write(self.imgsFn)
     
     def createOutputStep(self):
